Completed planet spawning method, slightly re-weighted planets.
Completed the method for calculating the type of planet to spawn as well as slightly redoing the weights assigned to the different planet types. Still need to finish reweighting them though.

Method should work fine, though it is slightly more hardcoded than it should be. Will likely fix that later.