Added basic map generation class, added new xmlLoader method.
Now supports the generation of a map complete with a 2D tile grid and dynamic random star generation.

Also slightly changed the xmlLoader to call itself from a base loadContent() method rather than loadXML(). This enables me to remove the folder declarations from the gameLoader class so that it doesn't have any potential interference as well as reusing the already declared folders in the xmlLoader class.

As a result, the gameLoader was updated to support the new map generation as well as xmlLoader structure.